Combine metadata and metadata db
