+++ metadata/reader_test.go
@@ -14,23 +14,43 @@
    limitations under the License.
 */
 
-package db
+/*
+   Copyright The containerd Authors.
+
+   Licensed under the Apache License, Version 2.0 (the "License");
+   you may not use this file except in compliance with the License.
+   You may obtain a copy of the License at
+
+       http://www.apache.org/licenses/LICENSE-2.0
+
+   Unless required by applicable law or agreed to in writing, software
+   distributed under the License is distributed on an "AS IS" BASIS,
+   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+   See the License for the specific language governing permissions and
+   limitations under the License.
+*/
+
+package metadata
 
 import (
 	"io"
 	"os"
+	"testing"
 
-	"github.com/awslabs/soci-snapshotter/metadata"
 	"github.com/awslabs/soci-snapshotter/ztoc"
 	bolt "go.etcd.io/bbolt"
 )
 
-func NewDbMetadataStore(sr *io.SectionReader, ztoc *ztoc.Ztoc, opts ...metadata.Option) (metadata.Reader, error) {
+func TestMetadataReader(t *testing.T) {
+	TestReader(t, newTestableReader)
+}
+
+func newTestableReader(sr *io.SectionReader, ztoc *ztoc.Ztoc, opts ...Option) (TestableReader, error) {
 	f, err := os.CreateTemp("", "readertestdb")
 	if err != nil {
 		return nil, err
 	}
-	defer f.Close()
+	defer os.Remove(f.Name())
 	db, err := bolt.Open(f.Name(), 0600, nil)
 	if err != nil {
 		return nil, err
@@ -39,8 +59,8 @@ func NewDbMetadataStore(sr *io.SectionReader, ztoc *ztoc.Ztoc, opts ...metadata.
 	if err != nil {
 		return nil, err
 	}
-	return &readCloser{
-		Reader: r,
+	return &testableReadCloser{
+		TestableReader: r.(*reader),
 		closeFn: func() error {
 			db.Close()
 			return os.Remove(f.Name())
@@ -48,12 +68,12 @@ func NewDbMetadataStore(sr *io.SectionReader, ztoc *ztoc.Ztoc, opts ...metadata.
 	}, nil
 }
 
-type readCloser struct {
-	metadata.Reader
+type testableReadCloser struct {
+	TestableReader
 	closeFn func() error
 }
 
-func (r *readCloser) Close() error {
+func (r *testableReadCloser) Close() error {
 	r.closeFn()
-	return r.Reader.Close()
+	return r.TestableReader.Close()
 }
