+++ soci/util_test.go
